Rename project from cram to prysk
