--- cram/__main__.py
@@ -1,10 +1,0 @@
-"""Main module (invoked by "python3 -m cram")"""
-
-import sys
-
-import cram
-
-try:
-    sys.exit(cram.main(sys.argv[1:]))
-except (BrokenPipeError, KeyboardInterrupt):
-    pass
