+++ prysk/__init__.py
@@ -1,6 +1,6 @@
 """Functional testing framework for command line applications"""
 
-from cram._main import main
-from cram._test import test, testfile
+from prysk._main import main
+from prysk._test import test, testfile
 
 __all__ = ['main', 'test', 'testfile']
