+++ prysk/__main__.py
@@ -1,0 +1,10 @@
+"""Main module (invoked by "python3 -m prysk")"""
+
+import sys
+
+import prysk
+
+try:
+    sys.exit(prysk.main(sys.argv[1:]))
+except (BrokenPipeError, KeyboardInterrupt):
+    pass
