+++ prysk/cli.py
@@ -1,0 +1,16 @@
+"""Main module (invoked by "python3 -m prysk")"""
+
+import sys
+
+import prysk
+
+
+def main():
+    try:
+        sys.exit(prysk.main(sys.argv[1:]))
+    except (BrokenPipeError, KeyboardInterrupt):
+        pass
+
+
+if __name__ == '__main__':
+    main()
