+++ tests/doctest.t
@@ -1,7 +1,7 @@
-Set up cram alias and example tests:
+Set up prysk alias and example tests:
 
   $ . "$TESTDIR"/setup.sh
 
 Run doctests:
 
-  $ doctest "$TESTDIR"/../cram
+  $ doctest "$TESTDIR"/../prysk
