+++ tests/md5.t
@@ -1,4 +1,4 @@
-Set up cram alias and example tests:
+Set up prysk alias and example tests:
 
   $ . "$TESTDIR"/setup.sh
