+++ tox.ini
@@ -1,6 +1,7 @@
 [tox]
 #TODO: adjust set of used python interpreter(s)
-envlist=py3-{integration}-{dash,bash,zsh},py3-{coverage}
+envlist=py3-{integration}-{dash,bash,zsh}
+#,py3-{coverage}
 
 [testenv]
 setenv =
@@ -13,7 +14,7 @@ deps =
     -r{toxinidir}/requirements.txt
 
 commands =
-    integration:    cram {env:TESTOPTS:--shell=dash} {toxinidir}/tests
+    integration:    prysk {env:TESTOPTS:--shell=dash} {toxinidir}/tests
 
 [testenv:coverage]
 setenv =
